requirements.txt
astroid==3.0.1
beautifulsoup4==4.12.2
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
dill==0.3.7
distlib==0.3.7
filelock==3.13.1
identify==2.5.32
idna==3.6
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
mysql-connector-python==8.2.0
nodeenv==1.8.0
packaging==23.2
platformdirs==4.0.0
pluggy==1.3.0
pre-commit==3.5.0
protobuf==4.21.12
pylint==3.0.2
pytest==7.4.3
PyYAML==6.0.1
requests==2.31.0
requests-mock==1.11.0
setuptools==69.0.2
six==1.16.0
soupsieve==2.5
tomlkit==0.12.3
urllib3==2.1.0
virtualenv==20.25.0
